+++ crates/ruma-client-api/CHANGELOG.md
@@ -15,6 +15,7 @@ Breaking changes:
 * Use `NewPushRule` to construct a `push::set_pushrule::v3::Request`
 * `Error` is now an enum because endpoint error construction is infallible (see changelog for
   `ruma-common`); the previous fields are in the `Standard` variant
+* Use `GlobalAccountDataEventType` for `event_type` in `config::get_global_account_data`
 
 Improvements: